requirements.txt
Flask==3.0.3
flask-cors
python-dotenv
psycopg2-binary
requests
openai
